requirement.txt
aioredis==1.3.1
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.5
channels==3.0.1
channels-redis==2.4.2
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
dj-database-url==0.5.0
Django==3.1.7
django-background-tasks==1.2.5
django-compat==1.0.15
django-heroku==0.3.1
djangorestframework==3.12.2
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
mouse==0.7.1
MouseInfo==0.1.3
msgpack==0.6.2
mss==6.1.0
Naked==0.1.31
oauthlib==3.1.1
paytmchecksum==1.7.0
Pillow==8.1.2
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycparser==2.20
pycryptodome==3.10.1
pygame==2.0.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.3
pyOpenSSL==20.0.1
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
PyTweening==1.0.3
pytz==2021.1
pywhatkit==4.7
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
service-identity==21.1.0
shellescape==3.8.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tqdm==4.61.0
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
urllib3==1.26.5
whitenoise==5.2.0
wikipedia==1.4.0
zope.interface==5.4.0
